requirements.txt
mnemonic
bip32utils
eth-account
requests
python-telegram-bot